Reverted to checking tags for block light sources
Added lava and lava bucket as a light source (still deadly though) because it makes sense.
Only check for fire_immunity on LivingEntities to prevent crashes.
Shortened block cache duration to be more responsive.
Increased block check radius from 2 to 3.
Blocks are now checked in a circle instead of a cube.